Fixed some warnings
git-svn-id: https://svn.aros.org/svn/aros/trunk@19314 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/libs/diskfont/bullet.c
@@ -500,7 +500,7 @@ struct AADiskFontHeader *OTAG_AllocFontStruct(STRPTR name, UWORD numchars, LONG 
 
 			dfh->dfh_FileID     = DFH_ID;
 			dfh->dfh_DF.ln_Name = dfh->dfh_Name;
-			dfh->dfh_Segment    = MAKE_REAL_SEGMENT(dfh);
+			dfh->dfh_Segment    = (LONG)MAKE_REAL_SEGMENT(dfh);
 
 			i = strlen(FilePart(name)) + 1;
 			if (i >= sizeof(dfh->dfh_Name)) i = sizeof(dfh->dfh_Name) - 1;
